[20]: Use go router builder with static routing
