+++ feature_modules/mensa/lib/src/pages/mensa_details_page.dart
@@ -1,149 +1,41 @@
-import 'package:core/components.dart';
-import 'package:core/constants.dart';
-import 'package:core/themes.dart';
-import 'package:flutter/material.dart';
+import 'package:core/routes.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mensa/src/bloc/mensa_cubit/mensa_menu_cubit.dart';
-import 'package:mensa/src/bloc/mensa_cubit/mensa_menu_state.dart';
 import 'package:mensa/src/repository/api/api.dart';
-import 'package:mensa/src/views/views.dart';
-import 'package:mensa/src/widgets/my_taste_button.dart';
+import 'package:mensa/src/repository/mensa_repository.dart';
+import 'package:mensa/src/utils/mensa_day.dart';
+
+import '../views/mensa_details_view.dart';
 
 class MensaDetailsPage extends StatelessWidget {
   const MensaDetailsPage({
+    required this.arguments,
     super.key,
-    required this.mensaModel,
   });
 
-  final MensaModel mensaModel;
+  final Object? arguments;
 
   @override
   Widget build(BuildContext context) {
-    final openingHours = mensaModel.openingHours;
+    final mensaModel = (arguments as MensaDetailsRouteArguments).mensaModel as MensaModel;
+    final MensaDay selectedMensaDay = (arguments as MensaDetailsRouteArguments).mensaDay as MensaDay;
 
-    return Scaffold(
-      backgroundColor: context.colors.neutralColors.backgroundColors.base,
-      body: CustomScrollView(
-        slivers: [
-          SliverAppBar(
-            expandedHeight: 210,
-            pinned: true,
-            elevation: 0,
-            surfaceTintColor: Colors.transparent,
-            backgroundColor: context.colors.neutralColors.backgroundColors.base,
-            flexibleSpace: FlexibleSpaceBar(
-              background: Image.network(
-                'https://www.byak.de/preiseundco-proxy/images/projekte/538/63a07e3ad2d28676dc35fbb7/b20352fdf91b24c98b9483d5f9083c6e.jpg?w=1260&h=648',
-                fit: BoxFit.cover,
-              ),
-            ),
-            leading: const SafeArea(
-              child: Padding(
-                padding: EdgeInsets.all(LmuSizes.medium),
-                child: _DetailsBackButton(),
-              ),
-            ),
-            actions: [
-              Padding(
-                padding: const EdgeInsets.only(right: LmuSizes.mediumLarge),
-                child: GestureDetector(
-                  onTap: () {
-                    LmuBottomSheet.show(
-                      context,
-                      title: "My Taste",
-                    );
-                  },
-                  child: MyTasteButton(background: context.colors.neutralColors.backgroundColors.base),
-                ),
-              ),
-            ],
-          ),
-          SliverToBoxAdapter(
-            child: Padding(
-              padding: const EdgeInsets.only(
-                left: LmuSizes.mediumLarge,
-                right: LmuSizes.mediumLarge,
-                top: LmuSizes.mediumSmall,
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.symmetric(
-                      vertical: LmuSizes.medium,
-                    ),
-                    child: LmuText.h1(
-                      mensaModel.name,
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(
-                      vertical: LmuSizes.medium,
-                    ),
-                    child: LmuText.body(
-                      mensaModel.location.address,
-                      color: context.colors.neutralColors.textColors.mediumColors.base,
-                    ),
-                  ),
-                  Divider(thickness: .5, height: 0, color: context.colors.neutralColors.borderColors.seperatorLight),
-                  LmuListDropdown(
-                    title: "Heute geöffnet bis ",
-                    titleColor: Colors.green,
-                    items: openingHours
-                        .map((e) => LmuListItem.base(
-                              title: e.day,
-                              hasVerticalPadding: false,
-                              hasHorizontalPadding: false,
-                              trailingTitle:
-                                  '${e.startTime.substring(0, e.startTime.length - 3)} - ${e.endTime.substring(0, e.endTime.length - 3)} Uhr',
-                            ))
-                        .toList(),
-                  ),
-                  const SizedBox(height: LmuSizes.medium),
-                  BlocBuilder<MensaMenuCubit, MensaMenuState>(
-                    builder: (context, state) {
-                      if (state is MensaMenuLoadInProgress) {
-                        return const MensaMenuLoadingView();
-                      } else if (state is MensaMenuLoadSuccess) {
-                        return MensaMenuContentView(
-                          mensaMenuModels: state.mensaMenuModels,
-                        );
-                      }
-                      return const MensaMenuErrorView();
-                    },
-                  ),
-                  const SizedBox(height: LmuSizes.medium),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
+    final mensaRepository = ConnectedMensaRepository(
+      mensaApiClient: MensaApiClient(),
     );
-  }
-}
 
-class _DetailsBackButton extends StatelessWidget {
-  const _DetailsBackButton();
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        Navigator.of(context).pop();
-      },
-      child: Container(
-        width: 36,
-        height: 36,
-        decoration: BoxDecoration(
-          shape: BoxShape.circle,
-          color: context.colors.neutralColors.backgroundColors.tile,
+    return BlocProvider<MensaMenuCubit>(
+      create: (context) => MensaMenuCubit(
+        mensaRepository: mensaRepository,
+      )..loadMensaMenuData(
+          mensaModel.canteenId,
+          selectedMensaDay.year,
+          30.toString(),
+          true,
         ),
-        child: LmuIcon(
-          icon: Icons.arrow_back,
-          size: LmuIconSizes.medium,
-          color: context.colors.neutralColors.textColors.strongColors.base,
-        ),
+      child: MensaDetailsView(
+        mensaModel: mensaModel,
       ),
     );
   }
