+++ feature_modules/mensa/lib/src/views/views.dart
@@ -1,6 +1,8 @@
 export 'mensa_content_view.dart';
+export 'mensa_details_view.dart';
 export 'mensa_error_view.dart';
 export 'mensa_loading_view.dart';
+export 'mensa_main_view.dart';
 export 'mensa_menu_content_view.dart';
 export 'mensa_menu_error_view.dart';
 export 'mensa_menu_loading_view.dart';
