+++ feature_modules/mensa/pubspec.yaml
@@ -25,7 +25,9 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^2.0.0
   json_serializable: ^6.6.2
-  build_runner: ^2.1.5
+  go_router_builder: ^2.7.1
+  build_runner: any
+  build_verify: any
 
 
 flutter:
