+++ feature_modules/settings/lib/settings.dart
@@ -1,1 +1,1 @@
-export 'src/routes/routes.dart';
+export 'src/routes/settings_routes.dart';
