--- feature_modules/settings/lib/src/routes/routes.dart
@@ -1,2 +1,0 @@
-export 'settings_appearance_route.dart';
-export 'settings_main_route.dart';
