--- feature_modules/settings/lib/src/routes/settings_appearance_route.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../pages/settings_appearance_page.dart';
-
-class SettingsApperanceRoute extends StatelessWidget {
-  const SettingsApperanceRoute({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const SettingsApperancePage();
-  }
-}
