--- feature_modules/settings/lib/src/routes/settings_main_route.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-import '../pages/pages.dart';
-
-class SettingsMainRoute extends StatelessWidget {
-  const SettingsMainRoute({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const SettingsMainPage();
-  }
-}
