+++ feature_modules/settings/pubspec.yaml
@@ -22,6 +22,9 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^3.0.0
+  go_router_builder: ^2.7.1
+  build_runner: any
+  build_verify: any
 
 
 flutter:
